// src/lib/api/model/filterEntryDtoTier.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Wildcat Radio v2 API
 * Control plane — see docs/final-build-plan/03-API-CONTRACT.md
 * OpenAPI spec version: 0.0.1
 */

export type FilterEntryDtoTier = typeof FilterEntryDtoTier[keyof typeof FilterEntryDtoTier];


export const FilterEntryDtoTier = {
  BLOCK: 'BLOCK',
  WATCH: 'WATCH',
} as const;